Create api architecture template
