+++ api/games/index.js
@@ -1,0 +1,14 @@
+const express = require('express')
+const router = express.Router()
+
+router.get('/dice/:n',(req,res) => {
+    const number = dice.dice(req.params.n)
+    const result = { result : number }
+    res.json(result)
+})
+
+router.get('/dice/:n/:emoticons', (req, res) => {
+
+})
+
+module.exports = router
